Allow string as input of key.

+++ src/aes256.jl
@@ -106,7 +106,7 @@ struct Aes256EncryptKey <: AbstractAesEncryptKey
     Aes256EncryptKey(keys::NTuple{15, __m128i}) = new(keys)
 end
 Aes256EncryptKey(key::ByteSeq) = Aes256EncryptKey(pad_or_trunc(key, Val(32)))
-Aes256EncryptKey(key::Integer) = Aes256EncryptKey(to_bytes(key))
+Aes256EncryptKey(key) = Aes256EncryptKey(to_bytes(key))
 Aes256EncryptKey(key1::UInt128, key2::UInt128) = Aes256EncryptKey((to_bytes(key1)..., to_bytes(key2)...))
 function Aes256EncryptKey(key::NTuple{32,UInt8})
     key1 = to_uint128(key[1:16]) |> to_m128i
